refactor(bill-core): 重构金额相关数据类型和验证规则- 将金额值的数据类型从 bigInteger 改为 decimal，以支持更精确的数值 - 修改金额汇总逻辑，将结果转换为字符串类型 - 更新 BillRequest 中的验证规则，移除对 currency 的 required 验证 -优化 Money 类的构造函数和 make 方法，支持多种数值类型输入

+++ database/migrations/2025_01_17_090109_create_bills_table.php
@@ -17,7 +17,7 @@ return new class extends Migration {
             $table->string('bill_category')->nullable();
             $table->timestamp('bill_time');
             $table->string('amount_currency')->default('CNY');
-            $table->bigInteger('amount_value')->default(0);
+            $table->decimal('amount_value')->default(0);
             $table->string('payee_type')->nullable();
             $table->string('payee_id')->nullable();
             $table->string('pay_method')->nullable();
